web/requirements.txt
google-cloud-pubsub
Flask==2.2.4